update architecture backend
